page builder drop old canvas
