+++ lib/shared/screens/admin/shared/components/page-builder/canvas/index.js
@@ -7,7 +7,7 @@ import React, {PropTypes} from 'react';
 import {bindActionCreators} from 'redux';
 import {connect} from 'react-redux';
 
-import Canvas from './new-canvas.jsx';
+import Canvas from './canvas.jsx';
 
 @connect(
   (state) => ({
